cleaned code implemented by Franco
